Reconcile Rails code with Traject loading
