+++ marc_to_solr/lib.rb
@@ -12,4 +12,3 @@ require_relative './lib/embargo_date_extractor'
 require_relative './lib/language_service'
 require_relative './lib/language_extractor'
 require_relative './lib/pul_solr_json_writer'
-require_relative './lib/mms_records_report'
