+++ marc_to_solr/lib/traject_config.rb
@@ -10,6 +10,7 @@ require 'stringex'
 require 'library_stdnums'
 require 'time'
 require 'iso-639'
+require_relative '../../app/models/mms_records_report'
 extend Traject::Macros::Marc21Semantics
 extend Traject::Macros::MarcFormats
 
@@ -121,9 +122,11 @@ end
 
 to_field 'figgy_1display' do |record, accumulator|
   mms_report = MmsRecordsReport.new.mms_records_report
-  next unless mms_report.keys.include?(record['001']&.value)
+  next unless mms_report.key?(record['001']&.value)
+
   open_items = mms_report[record['001'].value].select { |item| item['visibility']['label'] == 'open' }
-  next unless open_items.present?
+  next if open_items.blank?
+
   accumulator << open_items.to_json.to_s
 end
