+++ spec/services/alma/indexer_spec.rb
@@ -6,9 +6,10 @@ RSpec.describe Alma::Indexer, indexing: true, sidekiq: true do
   before do
     allow(ENV).to receive(:fetch).and_call_original
     allow(ENV).to receive(:fetch).with('CATALOG_SYNC_TOKEN').and_return('FAKE_TOKEN')
-    Rails.cache.clear
     stub_request(:get, 'https://figgy.princeton.edu/reports/mms_records.json?auth_token=FAKE_TOKEN')
-    .to_return(status: 200, body: File.open('spec/fixtures/files/figgy_report.json'))
+      .to_return(status: 200, body: File.open('spec/fixtures/files/figgy_report.json'))
+    # Create the cached report before Traject tries to get it within a sub-process
+    MmsRecordsReport.new.mms_records_report
   end
 
   describe '#index_file' do
